Fix cmake compile failing when including project as a library
When using FetchContent, the compilation fails because CMAKE_SOURCE_DIR is referring to the source dir of the project including the library, not to the source dir of this library. I changed CMAKE_SOURCE_DIR to CMAKE_CURRENT_SOURCE_DIR, which refers to the source dir of this library and the compilation works.